Music page is now fully gatsby!
