+++ package.json
@@ -20,6 +20,7 @@
     "fontfaceobserver": "^2.0.13",
     "gatsby-plugin-google-analytics": "^2.1.4",
     "gatsby-plugin-google-fonts": "^1.0.1",
+    "gatsby-plugin-layout": "^1.1.7",
     "gatsby-plugin-mdx": "^1.0.28",
     "gatsby-plugin-styled-components": "^3.1.2",
     "lodash": "^4.17.10",
